nerves_dht v1.1.0 NervesDHT

Elixir library to read the DHT series of humidity and temperature sensors on a Raspberry Pi.

The library expose direct access to the sensors via read/4 and stream/3.

If your application has multiple processes reading the same sensor concurrently you should add NervesDHT under your supervisor (see child_spec/1). It ensures that only one read operation at the time is executed. In case of multiple sources asking for a read while the operation is in progress, they will get the result of the ongoing read.

Types

delay()

delay() :: non_neg_integer

Delay between reading retries in [ms]

device_id()

device_id() :: atom

Device identifier

interval()

interval() :: non_neg_integer

Streaming period in [ms]

pin()

pin() :: 0..31

GPIO pin (using BCM numbering)

reason()

reason() :: :timeout | :checksum | :argument | :gpio

Error reason

result()

result ::
  {:ok, humidity :: float, temperature :: float} |
  {:error, reason}

Reading result

retries()

retries() :: non_neg_integer

Number of retries on transient errors

sensor()

sensor() :: :dht11 | :dht22 | :am2302

Sensor model

Functions

child_spec(list)

child_spec([name: device_id, sensor: sensor, pin: pin]) :: Supervisor.child_spec

Return the child specification to put the a named device under your supervisor tree. The device can be used to read concurrently from the sensor.

Add to you supervisor:

children = [
  {NervesDHT, [name: :my_sensor, sensor: :am2302, pin: 17]},
  ...
]
Supervisor.start_link(children, strategy: :one_for_one)

Read from the named :my_sensor device:

NervesDHT.device_read(:my_sensor)

device_read(device_id)

device_read(device_id) :: result

Read DHT sensor values from the named device device_id. The underlying read/4 operation will apply the default retries, delay strategy.

See child_spec/1, read/4.

device_stream(device_id, interval \\ 2000)

device_stream(device_id, interval) :: Enumerable.t

Return a Stream of sensor readings from the named device device_id.

See child_spec/1, stream/3.

read(sensor, pin, retries \\ 3, delay \\ 2000)

read(sensor, pin, retries, delay) :: result

Read DHT sensor values.

Read from the specified sensor type (DHT11, DHT22, or AM2302) on specified pin and return humidity (as a floating point value in percent) and temperature (as a floating point value in Celsius) as a tuple {:ok, humidity, temperature}.

Note that because the sensor requires strict timing to read and Linux is not a real time OS, a result is not guaranteed to be returned! The function will attempt to read multiple times (up to the specified max retries+1) until a good reading can be found. If a good reading cannot be found after the amount of retries the function returns {:error, :timeout}. The delay between retries is by default 2 seconds, but can be overridden (also note the DHT sensor cannot be read faster than about once every 2 seconds).

Examples

iex> NervesDHT.read(:am2302, 17)
{:ok, 55.1, 24.719}

stream(sensor, pin, interval \\ 2000)

stream(sensor, pin, interval) :: Enumerable.t

Return a Stream of sensor readings.

interval is the wait period beetwen two consecutive read attempts and defaults to 2 seconds. Since the device takes some x time to transmit the reading, the Stream will push data with a period >= (interval + x).

Examples

iex> NervesDHT.stream(:am2302, 17) |> Enum.take(2)
[{:ok, 55.1, 24.719}, {:ok, 55.12, 24.9}]